+++ src/recipes/mod.rs
@@ -1,2 +1,3 @@
 //! Extended ZooKeeper recipes from [Apache Curator](http://curator.apache.org/).
 pub mod cache;
+pub mod shared_lock;